Create README.md
This file describes the data files and R programs necessary to replicates tables and simulations for the paper “Dynamic Panel Threshold Spatial Durbin Model with Covariate-Dependent Thresholds”. The contents of the .zip file are detailed below. 

The R package “DPTSDCDT_1.0.5.tar.gz” contains R functions used to estimate the suggested model in this paper.

In the “estimation” folder, running the file “model.R” allows the replication of Table 1 of the paper (Mean and Std.dev). 

In the “power&size” folder, running the file “size1.R”, “power1.R”, “size2.R” and “power2.R” allows to allow the replication of Table 2 of the paper.